[ORB-00182] Remove retired orbit-design surface
