+++ crates/orbit-core/src/runtime/orbit_tool_host/mod.rs
@@ -1,6 +1,5 @@
 mod adr_tools;
 mod artifact_redaction;
-mod design_tools;
 mod dispatch;
 mod docs_tools;
 mod friction_tools;
