--- crates/orbit-tools/src/builtin/orbit/design/mod.rs
@@ -1,3 +1,0 @@
-pub mod init;
-pub mod list;
-pub mod show;
