+++ crates/orbit-tools/src/builtin/orbit/mod.rs
@@ -1,5 +1,4 @@
 pub mod adr;
-pub mod design;
 pub mod docs;
 pub mod duel;
 pub mod friction;
@@ -38,9 +37,6 @@ pub fn register(registry: &mut ToolRegistry) {
     registry.register(adr::show::OrbitAdrShowTool);
     registry.register(adr::supersede::OrbitAdrSupersedeTool);
     registry.register(adr::update::OrbitAdrUpdateTool);
-    registry.register(design::init::OrbitDesignInitTool);
-    registry.register(design::list::OrbitDesignListTool);
-    registry.register(design::show::OrbitDesignShowTool);
     registry.register(docs::OrbitDocsListTool);
     registry.register(docs::OrbitDocsShowTool);
     registry.register(docs::OrbitDocsSearchTool);
