+++ crates/orbit-tools/src/lib.rs
@@ -73,9 +73,6 @@ pub enum OrbitBuiltinAction {
     AdrList,
     AdrUpdate,
     AdrSupersede,
-    DesignInit,
-    DesignList,
-    DesignShow,
     DocsList,
     DocsShow,
     DocsSearch,
