+++ crates/orbit-tools/tests/public_tool_surface.rs
@@ -55,9 +55,6 @@ fn workflow_critical_tools_remain_registered() {
         "github.pr.review",
         "github.pr.review.comment",
         "github.pr.view",
-        "orbit.design.init",
-        "orbit.design.list",
-        "orbit.design.show",
         "orbit.graph.callers",
         "orbit.graph.deps",
         "orbit.graph.history",
